+++ generation/dnp3/src/main/scala/com/automatak/render/dnp3/enums/CommandResult.scala
@@ -7,7 +7,7 @@ object CommandResult {
 
   private val comments = List("Opendnp3 API enum used for differentiating cases when a command sequence fails without a response from the outstation")
 
-  def apply(): EnumModel = EnumModel("CommandResult", comments, EnumModel.UInt8, codes, codes.last, Base10)
+  def apply(): EnumModel = EnumModel("CommandResult", comments, EnumModel.Integer, codes, codes.last, Base10)
 
   private val codes = List(
        EnumValue("RESPONSE_OK", 0,  "A response was received from the outstation, check the CommandStatus enumeration"),
